+++ tests/heif_struct_test.rs
@@ -7,7 +7,9 @@ fn get_version() {
     let lib_heif = LibHeif::new();
     let version = lib_heif.version();
     assert!(version[0] >= 1);
-    if cfg!(feature = "v1_18") {
+    if cfg!(feature = "v1_19") {
+        assert!(version[1] >= 19);
+    } else if cfg!(feature = "v1_18") {
         assert!(version[1] >= 18);
     } else if cfg!(feature = "v1_17") {
         assert!(version[1] >= 17);